aistore review changes
